Update stack.c
There was a logical error in the code such that if we attempt to display for the best case then the all elements would be displayed, but when we talk about the worst case when the elements are poped after pushing then only remaining elements would be displayed. 
So this is the updated code for the same so that the memory of the pop element would be available for the next elements if pushed. 
Hope this finds  great .